+++ CHANGELOG.md
@@ -7,7 +7,11 @@
   eliminates sync failure in the form of `StopIteration` errors (#36, #38,
   #44).  Synchronization will now be faster and will not raise errors, but
   may also be less accurate occasionally.
+
 * Option `live_output` is now compatible with Jupyter kernels (#21).
+
+* Added command-line option `--live-output`.  This changes the default
+  `live_output` value for all sessions to `true` (#21).